requirements.txt
RPi.GPIO
rpi_ws281x
expiringdict
suds
python-vlc
gTTS